pyproject.toml
[tool.poetry]
name = "repl_nix_scrybe"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.4"
PyGithub = "^1.58.1"
python-dotenv = "^1.0.0"
requests = "^2.28.2"
Brainshop = "0.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
